The intensity of the led changes according to Bluetooth signal intensity
The value of RSSI is returned in hexadecimal values, a new function has been added in the program to convert these RSSI to 2's complement decimal values. The value of RSSI returned by the HC-05 ranges from -100 to 0. Here The signal at a value of "0" is more intense while at the other hand the signal at "-100" is less intense.
The RSSI 2's complement decimal values has been divided into 4 types, the first one ranges from 0 to -25, here the led glows at full intensity(225). Next is from the range -25 to -50 now the led glow with a lower intensity(60) and next with a signal strength ranging from -50 to -75 , here the led glow with much lower intensity(10). At last, the last range is -75 to -100.Now the led will glow at the lowest intensity(5). Note that the red led pin is used as analog pin.